order-service/requirements.txt
fastapi
uvicorn[standard]
kafka-python
prometheus-client
psycopg2-binary
